Weapon Selection not Updating Fix
We've fixed a bug caused by our partial (But not full) transition to enums for the WeaponSlotIndex. We've now replaced the instances of integers that still remained (As far as we're aware), and the graphics are now updating properly.

+++ NGP-Coursework/Assets/!_MainProject/Code/Scripts/GameplayObjects/ServerWeaponController.cs
@@ -41,9 +41,9 @@ namespace Gameplay.GameplayObjects.Character
                 Debug.Log(weaponAttachmentSlot.name + " is equipped in Slot " + weaponAttachmentSlot.SlotIndex + ". Weapon Data: " + weaponAttachmentSlot.GetComponentInChildren<Weapon>().WeaponData.name);
                 switch (weaponAttachmentSlot.SlotIndex)
                 {
-                    case (int)WeaponSlotIndex.Primary: _primaryWeapon = weaponAttachmentSlot.GetComponentInChildren<Weapon>(); break;
-                    case (int)WeaponSlotIndex.Secondary: _secondaryWeapon = weaponAttachmentSlot.GetComponentInChildren<Weapon>(); break;
-                    case (int)WeaponSlotIndex.Tertiary: _tertiaryWeapon = weaponAttachmentSlot.GetComponentInChildren<Weapon>(); break;
+                    case WeaponSlotIndex.Primary: _primaryWeapon = weaponAttachmentSlot.GetComponentInChildren<Weapon>(); break;
+                    case WeaponSlotIndex.Secondary: _secondaryWeapon = weaponAttachmentSlot.GetComponentInChildren<Weapon>(); break;
+                    case WeaponSlotIndex.Tertiary: _tertiaryWeapon = weaponAttachmentSlot.GetComponentInChildren<Weapon>(); break;
                 }
             }
